dom/l10n/tests/mochitest/document_l10n/test_docl10n_initialize_after_parse.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        xmlns:html="http://www.w3.org/1999/xhtml"
        title="Test lazy injecting l10n resource links in XHTML environment">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script type="application/javascript">
  <![CDATA[
  SimpleTest.waitForExplicitFinish();

  document.addEventListener("readystatechange", async function() {
    if (document.readyState != "complete") {
      return;
    }
    ok(!document.l10n, "l10n not initialized");
    MozXULElement.insertFTLIfNeeded("crashreporter/aboutcrashes.ftl");
    ok(document.querySelector("linkset link[rel=localization][href='crashreporter/aboutcrashes.ftl']"), "link exists");
    ok(document.l10n, "l10n initialized");
    await document.l10n.ready;

    ok(document.getElementById("main-desc").textContent.length > 0, "Text updated after init");

    info("Confirming a second call doesn't inject another link");
    MozXULElement.insertFTLIfNeeded("crashreporter/aboutcrashes.ftl");
    is(document.querySelectorAll("linkset link[rel=localization]").length, 1, "link exists");

    SimpleTest.finish();
  });
  ]]>
  </script>

  <description id="main-desc" data-l10n-id="crash-reports-title"/>
</window>
